requirements.txt
pandas
pytorch
pytorch-lightning
transformers
tokenizers
sklearn